Remove dependencies
Musty previously depended upon Spork and Testament.

The dependency on Spork was for its dedenting function. However, Janet
now does this automatically for long-strings.

The dependency on Testament was for testing during development. This
dependency is now put under the `:dev-dependencies` key in the call to
`declare-project` and a task is created to allow for it to be installed
using JPM.